+++ src/repo/mongoose/model/notification.ts
@@ -1,0 +1,46 @@
+import * as mongoose from 'mongoose';
+
+const safe = { j: true, w: 'majority', wtimeout: 10000 };
+
+/**
+ * 汎用通知スキーマ
+ */
+const schema = new mongoose.Schema(
+    {},
+    {
+        collection: 'notifications',
+        id: true,
+        read: 'primaryPreferred',
+        safe: safe,
+        strict: false,
+        useNestedStrict: true,
+        timestamps: {
+            createdAt: 'createdAt',
+            updatedAt: 'updatedAt'
+        },
+        toJSON: {
+            getters: true,
+            virtuals: true,
+            minimize: false,
+            versionKey: false
+        },
+        toObject: {
+            getters: true,
+            virtuals: true,
+            minimize: false,
+            versionKey: false
+        }
+    }
+);
+
+export default mongoose.model('Notification', schema)
+    .on(
+        'index',
+        // tslint:disable-next-line:no-single-line-block-comment
+        /* istanbul ignore next */
+        (error) => {
+            if (error !== undefined) {
+                console.error(error);
+            }
+        }
+    );