+++ src/repository.ts
@@ -4,6 +4,7 @@
  */
 import { MongoRepository as ActionRepo } from './repo/action';
 import { MongoRepository as GMONotificationRepo } from './repo/gmoNotification';
+import { MongoRepository as NotificationRepo } from './repo/notification';
 import { MongoRepository as SendGridEventRepo } from './repo/sendGridEvent';
 import { MongoRepository as TaskRepo } from './repo/task';
 import { MongoRepository as TelemetryRepo } from './repo/telemetry';
@@ -11,6 +12,7 @@ import { MongoRepository as TransactionRepo } from './repo/transaction';
 
 export class Action extends ActionRepo { }
 export class GMONotification extends GMONotificationRepo { }
+export class Notification extends NotificationRepo { }
 export class SendGridEvent extends SendGridEventRepo { }
 export class Task extends TaskRepo { }
 export class Telemetry extends TelemetryRepo { }